pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = ""
authors = ["Artem Holodyr <artemmholodyr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
psycopg2 = "^2.9.5"
environs = {extras = ["django"], version = "^9.5.0"}
django-allauth = "^0.52.0"
django-crispy-forms = "^2.0"
crispy-bootstrap5 = "^0.7"
pillow = "^9.4.0"
django-debug-toolbar = "^3.8.1"
django-widget-tweaks = "^1.4.12"
pypdf2 = "^3.0.1"
pdf2jpg = "^1.1"
pdf2image = "^1.16.3"
django-cleanup = "^7.0.0"
django-rest-framework = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
